Componentize the tablature block inputs. This helps to clean up the ExpandableTablatureBlock and to maintain seperation of concerns.
